+++ gameplay/enemy_input_controller.gd
@@ -1,0 +1,7 @@
+extends Node
+
+func _ready() -> void:
+	EventBus.enemy_turn_started.connect(func():
+		print("BUahahah >:(")
+		GameplayController.end_enemy_turn()
+	)